requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# Spark & Big Data
pyspark==3.5.0
findspark==2.0.1
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2

# NLP & Text Processing
nltk==3.8.1
spacy==3.7.2

# Image Processing
opencv-python-headless==4.9.0.80
# dlib==19.24.2
Pillow==10.2.0
pillow-heif==0.15.0

# Data Processing
pyreadstat==1.2.6
openpyxl==3.1.2
xlrd==2.0.1

# HTTP & Network
requests==2.31.0
urllib3==2.1.0

# Utilities
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
humps==0.2.2
regex==2023.12.25

# HDFS
hdfs[kerberos]==2.7.3